pyproject.toml
[project]
name = "fast-api-clean-architecture"
version = "0.1.0"
description = ""
authors = [
    {name = "Bishal Bhattarai",email = "bhattaraibishal50@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4.0"

dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "fastapi-utils (>=0.8.0,<0.9.0)",
    "typing-inspect (>=0.9.0,<0.10.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
